fastapi-app/requirements.txt
fastapi
uvicorn
pydantic
mysql-connector-python
sqlalchemy
python-dotenv
